+++ example/lib/network_image_detection.dart
@@ -4,12 +4,13 @@ import 'package:arkit_plugin/arkit_plugin.dart';
 import 'package:flutter/material.dart';
 import 'package:vector_math/vector_math_64.dart' as vector;
 
-class ImageDetectionPage extends StatefulWidget {
+class NetworkImageDetectionPage extends StatefulWidget {
   @override
-  _ImageDetectionPageState createState() => _ImageDetectionPageState();
+  _NetworkImageDetectionPageState createState() =>
+      _NetworkImageDetectionPageState();
 }
 
-class _ImageDetectionPageState extends State<ImageDetectionPage> {
+class _NetworkImageDetectionPageState extends State<NetworkImageDetectionPage> {
   ARKitController arkitController;
   Timer timer;
   bool anchorWasFound = false;
@@ -29,7 +30,13 @@ class _ImageDetectionPageState extends State<ImageDetectionPage> {
             fit: StackFit.expand,
             children: [
               ARKitSceneView(
-                detectionImagesGroupName: 'AR Resources',
+                detectionImages: const [
+                  ARKitReferenceImage(
+                    name:
+                        'https://upload.wikimedia.org/wikipedia/commons/thumb/0/02/OSIRIS_Mars_true_color.jpg/800px-OSIRIS_Mars_true_color.jpg',
+                    physicalWidth: 0.2,
+                  ),
+                ],
                 onARKitViewCreated: onARKitViewCreated,
               ),
               anchorWasFound
@@ -37,7 +44,7 @@ class _ImageDetectionPageState extends State<ImageDetectionPage> {
                   : Padding(
                       padding: const EdgeInsets.all(8.0),
                       child: Text(
-                        'Point the camera at the earth image from the article about Earth on Wikipedia.',
+                        'Point the camera at Mars photo from the article about Mars on Wikipedia.',
                         style: Theme.of(context)
                             .textTheme
                             .headline
@@ -60,7 +67,8 @@ class _ImageDetectionPageState extends State<ImageDetectionPage> {
 
       final material = ARKitMaterial(
         lightingModelName: ARKitLightingModel.lambert,
-        diffuse: ARKitMaterialProperty(image: 'earth.jpg'),
+        diffuse: ARKitMaterialProperty(
+            image: 'https://www.classe.cornell.edu/~seb/celestia/marsc-1k.jpg'),
       );
       final sphere = ARKitSphere(
         materials: [material],