+++ ios/Classes/FlutterArkit.m
@@ -43,6 +43,7 @@
 @property (strong) SceneViewDelegate* delegate;
 @property (readwrite) ARConfiguration *configuration;
 @property BOOL forceUserTapOnCenter;
+@property (nonatomic, copy, null_resettable) NSSet<ARReferenceImage *> *referenceImages;
 @end
 
 @implementation FlutterArkitController
@@ -91,6 +92,8 @@
       [self updateSingleProperty:call andResult:result];
   } else if ([[call method] isEqualToString:@"updateMaterials"]) {
       [self updateMaterials:call andResult:result];
+  } else if ([[call method] isEqualToString:@"performHitTest"]) {
+      [self performHitTest:call andResult:result];
 #if REQUIRE_TRUEDEPTH_API
   } else if ([[call method] isEqualToString:@"updateFaceGeometry"]) {
       [self updateFaceGeometry:call andResult:result];
@@ -156,9 +159,14 @@
         if (ARWorldTrackingConfiguration.isSupported) {
             ARWorldTrackingConfiguration* worldTrackingConfiguration = [ARWorldTrackingConfiguration new];
             worldTrackingConfiguration.planeDetection = self.planeDetection;
-            NSString* detectionImages = params[@"detectionImagesGroupName"];
-            if ([detectionImages isKindOfClass:[NSString class]]) {
-                worldTrackingConfiguration.detectionImages = [ARReferenceImage referenceImagesInGroupNamed:detectionImages bundle:nil];
+            NSString* detectionImagesGroupName = params[@"detectionImagesGroupName"];
+            if ([detectionImagesGroupName isKindOfClass:[NSString class]]) {
+                worldTrackingConfiguration.detectionImages = [self extractReferenceImagesFromGroup: detectionImagesGroupName];
+            }
+            NSArray<NSDictionary*>* detectionImages = params[@"detectionImages"];
+            if ([detectionImages isKindOfClass:[NSArray class]]) {
+                _referenceImages = [DecodableUtils parseARReferenceImagesSet:detectionImages];
+                worldTrackingConfiguration.detectionImages = _referenceImages;
             }
             _configuration = worldTrackingConfiguration;
         }
@@ -172,9 +180,14 @@
     } else if (configurationType == 2) {
         if (ARImageTrackingConfiguration.isSupported) {
             ARImageTrackingConfiguration* imageTrackingConfiguration = [ARImageTrackingConfiguration new];
-            NSString* trackingImages = params[@"trackingImagesGroupName"];
-            if ([trackingImages isKindOfClass:[NSString class]]) {
-                imageTrackingConfiguration.trackingImages = [ARReferenceImage referenceImagesInGroupNamed:trackingImages bundle:nil];
+            NSString* trackingImagesGroupName = params[@"trackingImagesGroupName"];
+            if ([trackingImagesGroupName isKindOfClass:[NSString class]]) {
+                imageTrackingConfiguration.trackingImages = [self extractReferenceImagesFromGroup: trackingImagesGroupName];
+            }
+            NSArray<NSDictionary*>* trackingImages = params[@"trackingImages"];
+            if ([trackingImages isKindOfClass:[NSArray class]]) {
+                _referenceImages = [DecodableUtils parseARReferenceImagesSet:trackingImages];
+                imageTrackingConfiguration.trackingImages = [self referenceImages];
             }
             _configuration = imageTrackingConfiguration;
         }
@@ -199,6 +212,10 @@
     SCNNode* node = [self.sceneView.scene.rootNode childNodeWithName:nodeName recursively:YES];
     [node removeFromParentNode];
     result(nil);
+}
+
+- (NSSet<ARReferenceImage *>*) extractReferenceImagesFromGroup:(NSString*) groupName {
+    return [ARReferenceImage referenceImagesInGroupNamed:groupName bundle:nil];
 }
 
 - (void)onGetNodeBoundingBox:(FlutterMethodCall*)call result:(FlutterResult)result {
@@ -234,20 +251,12 @@
         SCNNode *node = hitResults[0].node;
         [_channel invokeMethod: @"onNodeTap" arguments: node.name];
     }
-
-    NSArray<ARHitTestResult *> *arHitResults = [sceneView hitTest:touchLocation types:ARHitTestResultTypeFeaturePoint
-                                                + ARHitTestResultTypeEstimatedHorizontalPlane
-                                                + ARHitTestResultTypeEstimatedVerticalPlane
-                                                + ARHitTestResultTypeExistingPlane
-                                                + ARHitTestResultTypeExistingPlaneUsingExtent
-                                                + ARHitTestResultTypeExistingPlaneUsingGeometry
-                                                ];
-    if ([arHitResults count] != 0) {
-        NSMutableArray<NSDictionary*>* results = [NSMutableArray arrayWithCapacity:[arHitResults count]];
-        for (ARHitTestResult* r in arHitResults) {
-            [results addObject:[self getDictFromHitResult:r]];
-        }
-        [_channel invokeMethod: @"onARTap" arguments: results];
+    
+    NSArray<ARHitTestResult *> *arHitResults = [self getARHitResults:sceneView atLocation:touchLocation];
+    NSMutableArray<NSDictionary*>* resultsArray = [self ARHitResultsToArray:arHitResults];
+    
+    if(resultsArray != nil){
+        [_channel invokeMethod: @"onARTap" arguments: resultsArray];
     }
 }
 
@@ -413,6 +422,25 @@
     result(nil);
 }
 
+- (void) performHitTest:(FlutterMethodCall*)call andResult:(FlutterResult)result {
+    NSNumber* x = call.arguments[@"x"];
+    NSNumber* y = call.arguments[@"y"];
+    
+    double viewHeight = self.sceneView.bounds.size.height;
+    double viewWidth = self.sceneView.bounds.size.width;
+    
+    CGPoint location;
+    if ([x isKindOfClass:[NSNull class]] || [y isKindOfClass:[NSNull class]]){
+        location = self.sceneView.center;
+    }else{
+        location = CGPointMake(viewWidth * [x doubleValue], viewHeight * [y doubleValue]);
+    }
+    
+    NSArray<ARHitTestResult *> *arHitResults = [self getARHitResults:self.sceneView atLocation:location];
+    NSMutableArray<NSDictionary*>* resultsArray = [self ARHitResultsToArray:arHitResults];
+    result(resultsArray);
+}
+
 #pragma mark - Utils
 -(ARPlaneDetection) getPlaneFromNumber: (int) number {
   if (number == 0) {
@@ -575,6 +603,29 @@
         [dict setValue:[CodableUtils convertARAnchorToDictionary:result.anchor] forKey:@"anchor"];
     }
     return dict;
+}
+
+- (NSArray*) getARHitResults:(ARSCNView*)sceneView atLocation:(CGPoint)location{
+    NSArray<ARHitTestResult *> *arHitResults = [sceneView hitTest:location types:ARHitTestResultTypeFeaturePoint
+    + ARHitTestResultTypeEstimatedHorizontalPlane
+    + ARHitTestResultTypeEstimatedVerticalPlane
+    + ARHitTestResultTypeExistingPlane
+    + ARHitTestResultTypeExistingPlaneUsingExtent
+    + ARHitTestResultTypeExistingPlaneUsingGeometry
+    ];
+    return arHitResults;
+}
+
+- (NSMutableArray*) ARHitResultsToArray:(NSArray*)arHitResults{
+    if ([arHitResults count] != 0) {
+        NSMutableArray<NSDictionary*>* results = [NSMutableArray arrayWithCapacity:[arHitResults count]];
+        for (ARHitTestResult* r in arHitResults) {
+            [results addObject:[self getDictFromHitResult:r]];
+        }
+        return results;
+    }else{
+        return nil;
+    }
 }
 
 @end